Add `--latest-year` and `--latest-month` options
If both `--latest-year=<int>` and `--latest-month=<int>` are given, that year/month number will be used for the most recent year/month of data to include when generating reports. Otherwise, it'll default to the current year/month.

I also bumped the go version and a few dependency versions (most importantly cobra, to get `MarkFlagsRequiredTogether`, but I ended up bumping a few others while I was here).

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>